feat(osp): reduce duplication
