+++ src/components/pages/IDPManagement/IDPList.tsx
@@ -39,10 +39,45 @@ import {
   useEnableIDPMutation,
   useFetchIDPListQuery,
   useRemoveIDPMutation,
+  IDPProviderType,
 } from 'features/admin/idpApiSlice'
 
-export const IDPList = () => {
+const MenuItemOpenOverlay = ({
+  overlay,
+  id,
+  label,
+}: {
+  overlay: OVERLAYS
+  id: string
+  label: string
+}) => {
   const dispatch = useDispatch()
+
+  const openOverlay = async (
+    e: React.MouseEvent<HTMLElement, MouseEvent>,
+    overlay: OVERLAYS,
+    id: string
+  ) => {
+    try {
+      e.stopPropagation()
+      dispatch(show(overlay, id))
+    } catch (error) {
+      console.log(error)
+    }
+  }
+
+  return (
+    <MenuItem
+      onClick={(e: React.MouseEvent<HTMLElement, MouseEvent>) =>
+        openOverlay(e, overlay, id)
+      }
+    >
+      {label}
+    </MenuItem>
+  )
+}
+
+export const IDPList = ({ isOSP = false }: { isOSP?: boolean }) => {
   const { t } = useTranslation()
   const ti = useTranslation('idp').t
 
@@ -58,31 +93,7 @@ export const IDPList = () => {
   const [removeIDP] = useRemoveIDPMutation()
   const [enableIDP] = useEnableIDPMutation()
 
-  const doAddUsers = async (
-    e: React.MouseEvent<HTMLElement, MouseEvent>,
-    idp: IdentityProvider
-  ) => {
-    try {
-      e.stopPropagation()
-      dispatch(show(OVERLAYS.ADDUSERS_IDP, idp.identityProviderId))
-    } catch (error) {
-      console.log(error)
-    }
-  }
-
-  const doConfigure = async (
-    e: React.MouseEvent<HTMLElement, MouseEvent>,
-    idp: IdentityProvider
-  ) => {
-    try {
-      e.stopPropagation()
-      dispatch(show(OVERLAYS.UPDATE_IDP, idp.identityProviderId))
-    } catch (error) {
-      console.log(error)
-    }
-  }
-
-  const doConfirmDelete = async (
+  const doDelete = async (
     e: React.MouseEvent<HTMLElement, MouseEvent>,
     idp: IdentityProvider
   ) => {
@@ -127,70 +138,104 @@ export const IDPList = () => {
   }
 
   const renderMenu = (idp: IdentityProvider) => {
+    const menuItems = {
+      configure: (
+        <MenuItemOpenOverlay
+          overlay={OVERLAYS.UPDATE_IDP}
+          id={idp.identityProviderId}
+          label={ti('action.configure')}
+        />
+      ),
+      addUsers: (
+        <MenuItemOpenOverlay
+          overlay={OVERLAYS.ADDUSERS_IDP}
+          id={idp.identityProviderId}
+          label={ti('action.users')}
+        />
+      ),
+      delete: (
+        <MenuItem
+          onClick={(e: React.MouseEvent<HTMLElement, MouseEvent>) =>
+            !deleteLoading && doDelete(e, idp)
+          }
+          sx={{
+            color: deleteLoading ? '#b6b6b6' : '#111111',
+          }}
+          disabled={idp.enabled}
+        >
+          {ti('action.delete')}
+          {deleteLoading && (
+            <CircularProgress
+              size={15}
+              sx={{
+                marginLeft: '5px',
+              }}
+            />
+          )}
+        </MenuItem>
+      ),
+      register: (
+        <MenuItemOpenOverlay
+          overlay={OVERLAYS.REGISTER_OSP}
+          id={idp.identityProviderId}
+          label={ti('action.register')}
+        />
+      ),
+      registerNext: (
+        <MenuItemOpenOverlay
+          overlay={OVERLAYS.REGISTER_NEXT_OSP}
+          id={idp.identityProviderId}
+          label={ti('action.register_next')}
+        />
+      ),
+      consent: (
+        <MenuItemOpenOverlay
+          overlay={OVERLAYS.CONSENT_OSP}
+          id={idp.identityProviderId}
+          label={ti('action.consent')}
+        />
+      ),
+      enableToggle: (
+        <MenuItem
+          onClick={(e: React.MouseEvent<HTMLElement, MouseEvent>) =>
+            !disableLoading && doEnableDisableToggle(e, idp)
+          }
+          sx={{
+            color: disableLoading ? '#b6b6b6' : '#111111',
+          }}
+          disabled={
+            data &&
+            idp.enabled &&
+            data?.filter((idp: IdentityProvider) => idp.enabled).length < 2
+          }
+        >
+          {idp.enabled ? ti('action.disable') : ti('action.enable')}
+          {disableLoading && (
+            <CircularProgress
+              size={15}
+              sx={{
+                marginLeft: '5px',
+              }}
+            />
+          )}
+        </MenuItem>
+      ),
+    }
     return (
       <div className="action-menu">
         <DropdownMenu buttonText={ti('action.actions')}>
-          <MenuItem
-            onClick={(e: React.MouseEvent<HTMLElement, MouseEvent>) =>
-              doConfigure(e, idp)
-            }
-          >
-            {ti('action.configure')}
-          </MenuItem>
-          {idp.oidc?.clientId && (
-            <MenuItem
-              onClick={(e: React.MouseEvent<HTMLElement, MouseEvent>) =>
-                !disableLoading && doEnableDisableToggle(e, idp)
-              }
-              sx={{
-                color: disableLoading ? '#b6b6b6' : '#111111',
-              }}
-              disabled={
-                idpsData &&
-                idp.enabled &&
-                idpsData?.filter((idp: IdentityProvider) => idp.enabled)
-                  .length < 2
-              }
-            >
-              {idp.enabled ? ti('action.disable') : ti('action.enable')}
-              {disableLoading && (
-                <CircularProgress
-                  size={15}
-                  sx={{
-                    marginLeft: '5px',
-                  }}
-                />
-              )}
-            </MenuItem>
-          )}
-          {idp.enabled ? (
-            <MenuItem
-              onClick={(e: React.MouseEvent<HTMLElement, MouseEvent>) =>
-                doAddUsers(e, idp)
-              }
-            >
-              {ti('action.users')}
-            </MenuItem>
+          {menuItems.configure}
+          {isOSP ? (
+            <>
+              {idp.enabled && menuItems.register}
+              {idp.enabled && menuItems.registerNext}
+              {idp.enabled && menuItems.consent}
+            </>
           ) : (
-            <MenuItem
-              onClick={(e: React.MouseEvent<HTMLElement, MouseEvent>) =>
-                !deleteLoading && doConfirmDelete(e, idp)
-              }
-              sx={{
-                color: deleteLoading ? '#b6b6b6' : '#111111',
-              }}
-              disabled={idp.enabled}
-            >
-              {ti('action.delete')}
-              {deleteLoading && (
-                <CircularProgress
-                  size={15}
-                  sx={{
-                    marginLeft: '5px',
-                  }}
-                />
-              )}
-            </MenuItem>
+            <>
+              {idp.oidc?.clientId && menuItems.enableToggle}
+              {idp.enabled ? menuItems.addUsers : menuItems.delete}
+            </>
           )}
         </DropdownMenu>
       </div>
@@ -268,7 +313,15 @@ export const IDPList = () => {
           renderCell: ({ row }: { row: IdentityProvider }) => renderMenu(row),
         },
       ]}
-      rows={idpsData ?? []}
+      rows={(idpsData ?? [])
+        .filter(
+          (idp) =>
+            !isOSP ||
+            idp.identityProviderTypeId === (IDPProviderType.MANAGED as string)
+        )
+        .sort((a, b) =>
+          (a.displayName ?? '').localeCompare(b.displayName ?? '')
+        )}
       getRowId={(row: { [key: string]: string }) => row.identityProviderId}
       hasBorder={false}
     />
